backend/requirements.txt
# Faro backend – Python 3.12+
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
neo4j>=5.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pypdf>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
# Authentication & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
slowapi>=0.1.9
